+++ apps/frontend-pwa/src/components/learningElements/SingleQuestion.tsx
@@ -5,7 +5,7 @@ import {
   validateMcResponse,
   validateNumericalResponse,
   validateScResponse,
-} from '@lib/validateResponse'
+} from '@klicker-uzh/shared-components/src/utils/validateResponse'
 import { H3, UserNotification } from '@uzh-bf/design-system'
 import { useTranslations } from 'next-intl'
 import { useEffect, useState } from 'react'