+++ apps/frontend-pwa/src/components/liveSession/QuestionArea.tsx
@@ -1,4 +1,6 @@
 import { ElementDisplayMode, ElementType } from '@klicker-uzh/graphql/dist/ops'
+import StudentQuestion from '@klicker-uzh/shared-components/src/StudentQuestion'
+import { QUESTION_GROUPS } from '@klicker-uzh/shared-components/src/constants'
 import { push } from '@socialgouv/matomo-next'
 import { H2 } from '@uzh-bf/design-system'
 import dayjs from 'dayjs'
@@ -6,9 +8,6 @@ import localForage from 'localforage'
 import { useTranslations } from 'next-intl'
 import { without } from 'ramda'
 import React, { useEffect, useState } from 'react'
-
-import { QUESTION_GROUPS } from '@klicker-uzh/shared-components/src/constants'
-import StudentQuestion from '@klicker-uzh/shared-components/src/StudentQuestion'
 
 // TODO: notifications
 
@@ -144,7 +143,15 @@ function QuestionArea({
     type: string,
     instanceId: number
   ): void => {
-    if (value.length > 0 && QUESTION_GROUPS.CHOICES.includes(type)) {
+    if (type === ElementType.Kprim) {
+      handleNewResponse(
+        type,
+        instanceId,
+        Object.keys(value).flatMap<number[]>((key) =>
+          value[key] === true ? [parseInt(key)] : []
+        )
+      )
+    } else if (value.length > 0 && QUESTION_GROUPS.CHOICES.includes(type)) {
       handleNewResponse(type, instanceId, value)
     } else if (ElementType.FreeText === type) {
       handleNewResponse(type, instanceId, value)