+++ packages/shared-components/src/StudentQuestion.tsx
@@ -14,6 +14,11 @@ import KPAnswerOptions from './questions/KPAnswerOptions'
 import { NUMERICALAnswerOptions } from './questions/NUMERICALAnswerOptions'
 import { SCAnswerOptions } from './questions/SCAnswerOptions'
 import SessionProgress from './questions/SessionProgress'
+// TODO: merge validation logic in this file with the validateResponse utils
+import {
+  validateKprimResponse,
+  validateMcResponse,
+} from './utils/validateResponse'
 
 export interface StudentQuestionProps {
   activeIndex: number
@@ -84,25 +89,11 @@ export const StudentQuestion = ({
     (type: string): any =>
     (choice: any, selectedValue?: boolean): any =>
     (): void => {
-      const validateChoices = (newValue: any): boolean => {
-        switch (type) {
-          case ElementType.Sc:
-            return newValue.length === 1
-          case ElementType.Mc:
-            return newValue.length > 0
-          case ElementType.Kprim:
-            return (
-              Object.values(newValue).length === 4 &&
-              Object.values(newValue).filter((value) => value === undefined)
-                .length === 0
-            )
-        }
-      }
       if (inputValue && type === ElementType.Mc) {
         if (typeof inputValue === 'object') {
           setInputState({
             inputEmpty: false,
-            inputValid: validateChoices(inputValue),
+            inputValid: validateMcResponse(inputValue),
             inputValue: [],
           })
         }
@@ -112,7 +103,7 @@ export const StudentQuestion = ({
 
           return setInputState({
             inputEmpty: newInputValue.length === 0,
-            inputValid: validateChoices(newInputValue),
+            inputValid: validateMcResponse(newInputValue),
             inputValue: newInputValue,
           })
         }
@@ -121,7 +112,7 @@ export const StudentQuestion = ({
         const newInputValue = [...inputValue, choice]
         return setInputState({
           inputEmpty: false,
-          inputValid: validateChoices(newInputValue),
+          inputValid: validateMcResponse(newInputValue),
           inputValue: newInputValue,
         })
       }
@@ -134,7 +125,7 @@ export const StudentQuestion = ({
 
         return setInputState({
           inputEmpty: Object.keys(newInputValue).length === 0,
-          inputValid: validateChoices(choice),
+          inputValid: validateKprimResponse(newInputValue),
           inputValue: newInputValue,
         })
       }
@@ -234,13 +225,15 @@ export const StudentQuestion = ({
         onExpire={onExpire}
       />
 
-      <div
-        className={twMerge(
-          'mt-4 border-slate-300 flex-initial min-h-[6rem] bg-primary-10 border rounded leading-6 prose max-w-none prose-p:!m-0 prose-img:!m-0 p-4'
-        )}
-      >
-        <Markdown content={currentQuestion.content} />
-      </div>
+      {currentQuestion.content !== '<br>' && (
+        <div
+          className={twMerge(
+            'mt-4 border-slate-300 flex-initial min-h-[6rem] bg-primary-10 border rounded leading-6 prose max-w-none prose-p:!m-0 prose-img:!m-0 p-4'
+          )}
+        >
+          <Markdown content={currentQuestion.content} />
+        </div>
+      )}
 
       <div className="flex-1 mt-4">
         {typeof currentQuestion.type !== 'undefined' && (